# portal/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Karting Dashboard API",
        "description": "API for managing karting sessions, laps, drivers, and tracks",
        "contact": {
            "email": "support@karting-dashboard.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/drivers": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "List all drivers",
                "description": "Retrieve all drivers with their lap and session counts",
                "operationId": "558703a10247375755a77421384d8e38",
                "responses": {
                    "200": {
                        "description": "List of drivers"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Create a new driver",
                "description": "Create a new driver record",
                "operationId": "7f08894e03c5c4d6e18b957b1646c4f8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "nickname": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "color": {
                                        "type": "string",
                                        "maxLength": 7,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Driver created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/drivers/{id}": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Get driver details",
                "description": "Retrieve a driver with their laps and sessions",
                "operationId": "a31ab5afdce240854701ebd879eadc37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Driver details"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Driver not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Update a driver",
                "description": "Update an existing driver record",
                "operationId": "f39ca416707da5fd1421fdd820c9d0db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "nickname": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "color": {
                                        "type": "string",
                                        "maxLength": 7,
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Driver updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Driver not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Delete a driver",
                "description": "Remove a driver from the system",
                "operationId": "ff52ff62cc419864c51498a2a39672e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Driver deleted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Driver not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/stats/drivers": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Get driver statistics by account",
                "description": "Retrieve aggregated statistics for all drivers, grouped by user account",
                "operationId": "85a8432ff46862dcf910f82308356633",
                "responses": {
                    "200": {
                        "description": "Driver statistics"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}